ch(auth check files removed the useAuth and useForm files: replaced with redux api calls)
