+++ wire-transfer/frontend/core/helpers/utils/localStorageAPI.js
@@ -19,3 +19,41 @@ export const del = (key) => {
     window.localStorage.removeItem(key);
   }
 };
+
+
+
+
+export const loadInitialState = () =>{
+  if (typeof window !== "undefined") {
+  const localStorage =window.localStorage
+  const cachedUser = window.localStorage && JSON?.parse(localStorage.getItem("user"));
+  const cachedRole =
+    localStorage && JSON?.parse(localStorage.getItem("user_roles"));
+  const cachedToken = localStorage && localStorage.getItem("token");
+  const cachedAuthorization = localStorage.user && localStorage.token;
+
+  const initialState = {
+    isAuthenticated: cachedAuthorization ? true : false,
+    user: cachedUser ? cachedUser : null,
+    token: cachedToken ? cachedToken : null,
+    user_roles: cachedRole ? cachedRole : null,
+    error: null,
+    loading: false,
+    errFlag: false,
+    prevPath: "",
+  };
+
+  return initialState
+ }else{
+   return {
+    isAuthenticated:  false,
+    user:  null,
+    token:  null,
+    user_roles: null,
+    error: null,
+    loading: false,
+    errFlag: false,
+    prevPath: "",
+  }; 
+ }
+}
