+++ wire-transfer/frontend/package.json
@@ -15,10 +15,12 @@
     "next": "latest",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-hot-toast": "^2.2.0",
     "react-icons": "^4.2.0",
     "react-intersection-observer": "^8.32.0",
     "react-redux": "^7.2.6",
     "redux": "^4.1.2",
+    "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
