+++ wire-transfer/frontend/pages/_app.js
@@ -1,13 +1,22 @@
-
 import "../core/assets/styles/globals.css"
-import "../core/assets/styles/main.scss"
+import "../core/assets/styles/main.css"
 import React from "react"
 import { ChakraProvider } from "@chakra-ui/react";
+import { Provider as Provisioner }  from "react-redux";
+import QReduxStore from "../core/redux/store";
+
 export default function App({ Component, pageProps }) {
     return (
-    <ChakraProvider>
-      <Component {...pageProps} />
-    </ChakraProvider>
+
+   <>
+     {/*REDUX STATE ENHANCER*/}
+    <Provisioner store={QReduxStore}>
+          <ChakraProvider>
+                  <Component {...pageProps} />
+            </ChakraProvider>
+    </Provisioner>
+  </>
+    
     )
   }
